pyproject.toml
[tool.poetry]
name = "gmail-yaml-filters"
version = "0.10.0"
description = "A quick tool for generating Gmail filters from YAML rules."
authors = ["Alex Levy <mesozoic@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/mesozoic/gmail-yaml-filters"
homepage = "https://github.com/mesozoic/gmail-yaml-filters"
classifiers = [
  "Topic :: Communications :: Email :: Filters",
]

[tool.poetry.dependencies]
PyYAML = "^6.0"
google-api-python-client = "^2.40.0"
google-auth-oauthlib = "^0.5.2"
lxml = "^4.8.0"
python = "^3.6"

[tool.poetry.scripts]
gmail-yaml-filters = "gmail_yaml_filters.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
